Added 'show principal' and 'permission' commands to manipulate the current entitlement system.
Also added a bootstrap to auth.py - the builder of a database will
be added as an aqd_admin.

The basic entitlements are now enforced.  Anyone can 'show'.  Only
an aqd_admin can 'permission'.  A user needs to be both authenticated
and in a group besides 'nobody' in order to do anything else.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 638608]