Something of a large change here, still some rough edges.
Added rudimentary unit tests for end-to-end tests of the commands.
This is a rough translation of a few scripts I had, but formalized.

To run the tests, use the runtests.py script.  It will use 
unittest.conf in the same directory to build a new dev server and
run the tests.

All tests can be "independently", for varying values of independently.
For iterative testing, the full runtests.py can be run once, and
then (for example) the broker tests can be run by executing
broker/orderedsuite.py without needing to rebuild the db.

As a side-effect to this work, the broker now has something closer
to a proper configuration file.  I even managed to figure out how
to dynamically load modules so that the config file can be 
passed in on the command line.  Some minor updates to the wiki with
this information is pending.

A couple of the service-related unit tests currently fail.  I don't
see a point in tracking those down right now since that code is
going to change first thing next week.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 640107]