Added unit tests for the 'permission' and 'show principal' commands. (These tests do not cover entitlements - they simply check that new user principals can be added to the database and that their roles can be changed.)
Fixed the use of the UserPrincipal constructor by the broker to
pass in principal name as a keyword argument.

Added Yet Another Missing Comments Field workaround, this time for
the UserPrincipal table.

Fixed another unrelated bug in the unit test framework - for initial
database creation, a local copy of template-king should be available.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 641064]