First the little stuff: depends uses ipython 0.8.2 (latest). db.py is mostly cosmetic, but its also starting to use stuff from aquilon.config. It's really my damn OCD kicking in again since the big addition with this checkin means db.py is on its way out...
So the big news are the 2 new files debug and db_factory. debug is a small set of libraries for me which are just some short cuts for my style of in __main__ context unit tests. 

db_factory is how we will get meta, engine and session in the near future, though the code isn't 100% complete yet. I wanted to get this done because it contains a mock_engine object which takes an optional kw arg "file" to write it out otherwise straight to stderr with the ddl. if you create a custom meta outside the dbf you can use it to be bound at runtime and we'll be developing it some more for use in schema testing, archiving, diff'ing, and migrating. I'd love for nick and wes to code review it for me because i've been staring at it too long to be objective at this point.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 641425]