Don't quit right away when stdout is closed in the client
If stdout is closed before all error messages are received on
stderr then the error messages were not printed on the aq
output. Now we properly wait for the process to exit and the
only thing we do when stdout is closed is closing stdin so
that we avoid locking up.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 644292]