Populate command line option parser only with the options of the given subcommand
Previously, the option parser knew about all possible options
for all subcommands and thus you did not get an error when
specifying an option for the inappropriate command (like in
'aq show chassis --chassis NAME')

This changeset also changes restricts the position of the
subcommand arguments on the command line. Previously, it was
possible to run 'aq --aqhost quattorsrv status'. With the new
rules in place the subcommand has to precede all other options
on the command line, so now you have to run 'aq status
--aqhost quattorsrv' instead.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 644611]