Fixed up the dsdb callout code to work again.  However, temporarily disabled while the aqdb refactor is going on.  It needs to be re- enabled by uncommenting the two lines in add_host.py and del_host.py. Also, the logic needs to be written for update_interface.
Of note is that this change required an update to the init script.
The call to prod dsdb requires afs tokens, so the krun command
needed to be updated.

Need to decide how the unit tests are going to work.  Probably against
the test dsdb instance, but do they need to choose values to not
clash with others running the unit tests?

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 645572]