Dynamically determine the version of the running broker.
Modified status to check the config file.  If a config value for
the version has not been set (the default), it will attempt to
parse the file path of the source code.  If that fails, git describe
is run to determine the latest tag.  If that fails, the string "Unknown"
is used.