Adding ChassisSlot as hardware to track slot information.
The add_machine and show_machine commands can now use --slot and
the add_aurora_host command will create slot information if it can.

For now, Chassis does not auto-create slots, as it will need to be
modified later when it has associated hardware.

It turns out that putting cascade='delete' in the *backref* is
What We Want, at least for data like slot.