update_rack: Add this new command.
The input.xml change has been sitting in my queue since Aug 20.  This
command implements enough to fix row and column.  It also rewrites
plenary machine templates as necessary.

It introduces an undocumented feature - if we need to flush the
plenary machine files for a rack it can be done by issuing an update
rack and giving none of the optional arguments.