Upgrade to twisted 8.1.0 for both client and server.
Include fixes along the way for the unit tests to be able to
work correctly if the client is picking up packages from dev.

The first feature enhancement we're picking up is the ability to
integrate with standard Python logging, which sqlalchemy uses.  The
sqlalchemy logs can be enabled through the broker's config file,
and they will appear in aqd.log.