AQUILONAQD-136: Fixed incorrect network parameters in templates.
Many moons ago (at least two major reorgs ago) I had hard-coded that
templates would assume a /25 as the network.  Turns out that this was
never fixed after the network information was added to the interface
table.

It has been magically working all this time because dhcp was supplying
the correct values, which were overriding the ones in the templates.
This surfaced while attempting to configure a static interface.