Removed the use of log.err() from the code.
It looks like something in the way that logging is now configured has
an issue with it.  This isn't much of a problem - I switched away from
using log.err() a long time ago in favor of log.msg(), since log.err()
wasn't doing what I thought it would (and wasn't adding value).