Added an update machine command.
This allows for updating the chassis slot information and/or serial
number after the hardware entry has been created.  Most of the
hardware attributes can be changed, but those two are the only ones
tested by the unit tests.

The update machine command can take a --hostname parameter instead
of a --machine parameter.  We may also need to enhance it to take IP
addresses (or maybe even mac addresses) for both machine and chassis,
but we'll cross that bridge when it comes.