Another iteration of fixes for the new schema.  Most unit tests pass.
Of note:
- Removed name from chassis_manager and tor_switch, these will rely
  solely on a_name for now.
- Removed the polymorphic nature of Interface.  Possibly temporary.
- The interface table was brought more in line with current aqdb standards.
- The mac_address table was removed.

The two tests that fail involve show hostiplist.  Will wait on fixing
those until System and Host are updated.  There are also still plenty
of unit tests that need to be written.

Remove the HostList abstraction and give ServiceInstance a list of Systems.

This puts the name of the ServiceInstance directly in the table, and
enforces uniqueness over the service/instance combo instead of just a
unique name for every HostList.

To do this, a ServiceInstanceServer class was introduced to supply
the ordered list of Systems for the ServiceInstance.

fix repeat unique constraint name, add indexes on mac and ip