All unit tests pass.
I had to cripple and/or remove a few of the ones around interface and
ip changes... that whole system needs to be reworked.  Also, we've
temporarily lost the ability to have a static ip defined on eth1.

We're also still missing unit tests for some major areas like bind
server.