Show service by service, client, server all works, can qualify by instance
show_service and show_service_service both changed to allow services (or service instances) to be listed by naming the service, or by naming a client or a server.  Can supply --instance with any of those things, and it will modify the query.  When requesting lists like that (ie by client or server), it will just return nothing rather than error when there are no results.