Broke add interface into different commands for each of the hardware types.
Also added a convenience method for --hostname to redirect to --machine.

The --tor_switch and --chassis interfaces restrict to having a single
interface with an ip.  This simplifies the logic a bit and will probably
need to be revisited.

Simplified ip and mac addresses in use by the unit tests.

Successful test adding bmc interface to a machine.  Restricted the
machine plenary template to only include public interfaces - not sure
if that is the right way to go.