First pass at a show manager --missing command.
The command outputs a series of aq commands that could be used to
feed data back in to aqd with new managers.  The output is currently
broken in that the --ip option is the only way to add managers at the
moment.  Instead of outputing valid IP addresses, the string 'IP' is
used.

There is also a --format=csv option that just outputs the hostnames.