Updates to add interface to account for blindly built Dell machines.
This adds a lot of complication to the add_interface command, but
it should only kick in for the case of a management interface being
added where the mac address is already taken by a host and machine
that were created blindly.

Many unit tests were added, but almost none of the failure cases
are being tested.