Skip flushing host plenary templates that are incomplete.
Addresses-Issue: Jira/AQUILONAQD-142
This fix is inefficient.  Any templates missing an OS or personality
parameter throw an IncompleteError.  The current loops that flush
templates look for this and skip those hosts.  A better solution might
be to check for valid/built hosts and only attempt those.

The inefficieny could be reduced by skipping aurora hosts in the
relevant loops.