New update domain command to change the compiler in use by a domain.
It also allows the comments and owner of a domain to be updated.
This command is restricted to aqd_admin for now.
Updated the call to 'make' for a domain (also used by single-host
compiles) to use the domain compiler.

The config value for panc was renamed, and now specifies the value
used when a new domain is added.