add_manager: Fix to allow dynamic ip parameters to work correctly.
Addresses-Issue: Jira/AQUILONAQD-163
When the dynamic ip parameters were added to the commands beyond
add_host that took --ip, the required_parameters array for
add_manager_manager was not updated.  The command thought that
the --ip parameter was required when it no longer is (since the
ip can now be chosen via the dynamic parameters).