search_hardware: Fixed error on display of unnamed hardware.
Addresses-Issue: Jira/AQUILONAQD-168
There should not be unnamed hardware entities (hardware entities
both without a name and without an attached system).  However,
the code was set to handle this case and the handler contained
a think-o.  (Can't delegate to super-class property like you would a
method... because it's a property, not a method.)