Fixed traceback information for installed python source code.
When compiling, source files were being tagged as from /ms/dist
instead of /ms/dev.  That part is correct (we want tracebacks from
running processes to show /ms/dist), but the exact replacement
was not.  This fixes the replacement to point at the correct
location, including adding PROJ into the path and removing install.