Display list of available commands in multiple columns
Since we got a fair number of aq commands the output of 'aq help' is
way too long to fit on a single screen and most of the screen-estate
is used by the list of commands. This change changes the output to
print the command list in multiple columns which is much more visually
pleasing than the current single column list.