Removed use of decorators in broker commands.
A similar mechanism is still used, but done in a single method that
can take more complex actions.  The functionality of the decorators
is still exposed but with class-level flags.

Additionally, several defaults are now set in the code, such as
having the show and search commands automatically use a formatter.