Add code coverage option to the broker.
Invoking the broker with the new option causes a code coverage
report (courtesy the coverage module) to be written into the
requested file.

Added a --coverage option to runtests.py that causes a coverage
report for the broker to be written to the logdir as aqd.coverage.
This is not on by default as it increases the time for the tests
to run by more than a factor of 2.