Added use of ms.modulecmd for required broker modules.
This allows the broker to pull in the modules it needs directly instead
of having the information hard-coded in a wrapper script.  The module
versions are configurable in aqd.conf as the modules are added to the
environment before importing the bulk of aqd/aqdb.

Updated the build_db.py script to pull in a module for the database
if needed.