clean up the locks and exception handling for the non-flushed compiles
Now that we're not flushing everything when we compile, we need
to be smarter when making new hosts (or modifying them) in order
to ensure that the template gets written out correctly (and while I
was there, I finally implemented the logic that ensures that the template
files get reverted if the make_aquilon fails, ensuring that a "compile"
should always work). The locking is horrible API, but it works, and Wes
is fixing it all to a rosier world, so I'll just pretend that everything
is fine. And then I'll see if he reads these commit messages.