When moving the host to a different domain, allow for incomplete templates
A host may not be in a useful state to build (e.g. it's in the middle
of being defined). That shouldn't stop us from moving the host to
a different sandbox (i.e. "aq manage"), however to do that we have
to ensure that we trap IncompleteError exceptions from the template
writing code.