fix up the tests to work with new constraint
While changing the unbind client logic to ensure the
plenary templates were up-to-date, I've added a constraint
to say that you're not permitted to unbind a required service.
After a day of beating my head against the confusion that is
the testsuite, I've now re-organized things to (a) test this
new constraint and (b) not break because of this constraint.

  Ran 621 tests in 756.832s
  OK