Stop adding host msg for every client in a service instance.
Currently we're adding a host msg for every client in a service instance, when we get the -proto.  We aren't using that functionality for anything, and it takes a long, long time.  We can add the code later if we find that we need it, but with a flag (--hosts or something).