fix the makefile, which was no longer tracking dependencies
The makefile assumed that cwd was $QROOT and so that meant that
things like mkdep could use anchored regex to convert pathnames.
However, we're now using absolute pathnames and do not depend on
cwd being correct, so the anchored regex was no longer useful and
the dependencies were wrong. So, I've changed the anchored regex
to just look for /deps/ (i.e. a directory) instead of '^deps'.
While tracking down this problem, we decided to add some more
comments into the makefile attempting to explain what's happening
and furthermore, I've removed the cleandeps incantation, since it
was causing deps to always be blown away due to some previous changes
in our include path (we include the "hosts" directory, which always
changes, and therefore (according to cleandeps) always invalidates
all of the dependencies). The only reason for cleandeps is to ensure
that the deps are correct between moving a host between domains, but
we'll fix that separately by ensuring we clean up during an
"aq manage" and "aq del host".