fix the dot output to produce data amenable for import to other tools
The initial dot style that this thing was generating was basically
a hack: the boxes were done by making a "text" node and the text
was a slew of physical HTML formatting to make it look boxey. I've
made this use standard dot "record" type instead, so that it's more
amenable to importing (things like OmniGraffle can now read the dot
file happily). I might spend some time later doing more dot
enhancements, but this gets it going at least.