show_active_commands: Fix to work in all timezones.
The show_active_commands logic currently parses the log file to get its
information.  The parse has a quick regex for matching log lines that
includes the timestamp at the beginning of the line.  This regex only
accounted for timezones earlier than GMT (-0500) and not at or later
than GMT (+0000).