write host template when we change service binding, cleanup API
When running bind_client, we weren't flushing the host template,
which is kind of critical. Also, clean up the old service instance
plenary for the case that we're rebinding.
I've started to cleanup the api for templates: the templates "know"
which directory to use, so the calling code shouldn't need to look
it up.