A lot of fixes, but not very exciting
I've changed the API for the plenary templates so that you no
longer pass in the directory in which to write the files (unless
you really want to). I.e. all the "what-directory-does-this-thing-live"
is now contained in the templates code instead of everything looking
it up all over the place and passing it in. So, this change touches
many many files. The more interesting fix is in bind_client,
and make_aquilon, where it now creates the correct service instance
plenary templates.