add_interface: Do better plenary cleanup for blind builds.
Fix the Host plenary template and remove stale Machine plenary
templates when building machines blindly with a primary interface
and a bmc sharing a switch port.

This could be better, in particular most of the del_host routines
should be used if deleting a host.  This is still better than
what we had before.