Remove code that has never worked from add_*_host.py.
There was code testing to see if an argument was passed in, but it
did so by checking for the argument as a key in the argument list.
However, the argument is always present as a key and will be set
to None if the user did not supply it.

Therefore, the test was pointless.  Just removed the code.