add in archetype manipulation, allowing for non-aquilon compiles
Add in add/del/update archetype and modify the show command
to indicate which archetypes are compilable (and use --all to follow
the same conventions as other commands). I've then added a new "make"
command (and left "make aquilon" as-is for now).

map_service has also been modified to not require an archetype
when defining services which are not archetype based.