Allow unit tests to override template-king source.
The unit tests generally sync template-king from the broker
before starting.  Previously the broker used was configurable.
Now the whole path can be overridden to a different repository
on the broker or a local repository.  This makes it easy to
test against changes that have not yet been introduced.

Anyone using a custom conf file for unittests will need to
update it after applying this change.