Introduce add_license.py to fix python file headers.
The script is currently meant to be called manually.  It recurses
through subdirectories (by default starting in the project root)
and adjusts the ex/emacs line, the copyright notice, and the
license/warranty snippet.  It also cleans out older versions of
the same found elsewhere in the file.

The results need to be committed.