Added InternalError as a named exception.
This exception can be used to explicitly mark problems that are due
to programmer error rather than user error or external problems.

Any exception outside of a limited set within the broker will be shown
to the user as a 500 (Internal Error) - this just marks that behavior
when it is intended in the code.