there was a test for machine types no longer needed
When adding machines, it checks that the machine type is
reasonable. Previously, you could get non-reasonable machines
constructed and so that test was replicated where the machine
was used (in del_machine, and in add_host to machine). Because
this was replicated code, it was out-of-sync in the different
places so only some of the copies supported virtual machines. I've
just nuked the replicas and now we trust the code in add_machine.