fixed up (to some initial degree) the plenary templates for clusters
I've got the plenary templates for cluster into a working state.
Cluster clientdata now includes machines directly into the cluster
information (the host on the virtual machine may not yet be built,
so the hostname is no good, however the vm must be provisioned).
Modified add/del machine to keep the plenary in-sync (modifications
to the machine do not need to change the plenary, since the cluster
plenary simply does an include of the relevant machine tpl).

Still some cleanup of the cluster templates required: notably
the metacluster templates are still stubs with no real use. Until
we need them, that'll do fine.