cleanup of the template indexing code and the plenary templates
I've modified the functionality of the cluster templates to
produce profiles for clusters. In readiness for how this then
breaks downstream (non-host profiles), I've moved the template
indexing code into the templates module, prior to breaking the
build_index/notification functionality.

While working on this, I've cleaned up some of the other templates
by adding in more docstrings on what the templates are meant to
be used for and removing some duplicated code.