make cluster plenaries compilable
This is a fair size commit, please look carefully before merging!
I've made it so that cluster objects can be compiled. This neccessitates
the cluster being in a template domain, so there's a minor addition
for that to the aqdb model. This then changes the add cluster commands
to add in the domain argument, and a new "manage cluster" command
to change the domain. I've put a restriction into "manage host" such that
hosts in a cluster cannot be managed into different domains - only the
cluster can be managed into a different domain (and by extension, all hosts
within the cluster).

Once that's there, the plenaries themselves have been cleaned up and
made to work in this style: the cluster plenary is far simpler since it's
more like a host now and not a typical plenary that wants overrides.

Finally, the compilation is changed to cope with making new object
profiles. I've based this off wes's work on using the new compile system
instead of the old gnumakefile-based system.
I've decided not to make namespaced profiles, but instead to
have a completely alternate root of profiles (so, there's  output of
profiles/<hostname>.xml and there's clusters/<cluster>.xml). This means
each root (profiles and clusters) has its own profiles-info.xml with
associated code-fixes to allow this. This is easy, but wes may want
to review my very simplistic approach...

All tests pass of things I've touched, however I can't run the full
test-suite since some of the add interface tests late in the suite are
just hanging for me :(. But that's after all the compiling and checking
of both hosts and clusters.