notifications/indexing of profiles now supports namespaces
The build_index (and notifications off the index) now can cope
with "namespaces" of profiles (a.k.a. profiles-in-subdirectories).
slightly cleaned up log messages while I was there, so that it now
only logs the total number of notifications that were sent, instead
of a log message for each line (although errors are still logged).