move the template checkpoint/restore logic into templates module
Previously, the "make" command was grabbing the previous
template state before trying to make the host... and if the make
failed, was restoring the previous state. This doesn't work when
the templates are possibly multiple files, so I've moved this
logic into the template base class (as stash() and restore_stash()).
Also fixed the mtime of files when the stash is restored.