bugfix for recent commit, plus new PlenaryCollections
in a recent commit the template rewind logic (during a reconfigure,
try to change the template - if it fails to compile, put back the
previous version) was broken in that a rewind was just removing
the file, rather than putting back the old version. The reason
being that the "make" command was using a different variable
to unstash from, compared to where it had stashed the old content.
Yay for testsuite.

Added in new PlenaryCollection class to collect a bunch of plenary
templates behind a single facade. Switched PlenaryHost and
PlenaryCluster to use the PlenaryCollection facade.

Removed the refresh_cluster_plenaries() call since it's no longer
needed, and therefore touched all the commands that used to use
that function