some minimal support for nas disks
Provide share parameter to add_disk in order to specify
NasDisk objects. Modify the output (raw and template form)
appropriately.

We want the nasdisks to store both a name (e.g. sda) and the
scsi bus address. Therefore, using just the name is no good. So,
added an "address" field into the NasDisk type and modified the
add_disk to require that field for nas disks. The templates still
dump out the address within the disk specs for each virtual machine,
included within the cluster profile.