Add helpers for regression tests to match regexes in output.
This was needed to do multi-line matches in output.  Probably should
have been added a long time ago.  The new methods are searchoutput()
for enforcing positive matches and searchclean() for negative.

The searchoutput() helper will return the match object for use of
any of the group matches in further tests.