Fixed transient error in show_personality test.
Turns out there were two bugs here, and two wrongs sometimes made a
right.  (It certainly wasn't three lefts as then the test always
would have passed.)

There was a bug in the protobuf formatter for personality thresholds
of zero and a bug in some of the logic being used to test the code.
Depending on the phase of the moon (and the output order of the domain
list) the test could pass before either bug was triggered.  Fun stuff.