Complete revamp of error handling for commands.
The largest amount of work was done around making sure that writes
of plenary templates can be reversed.  The PlenaryCollection object
was enhanced to allow it to act as a generic wrapper for dealing
with errors in the simple cases.

Additionally all writes now force a refresh of the plenary's
database object.  This now requires that the session be flushed
before attempting a write.

Plenary.read() had a bug that checked the wrong location that was
fixed as part of this work.  When Plenary.read() was changed to use
the pathname() helper its use of read_file() was broken.  The
pathname() method returns a full path whereas read_file() expects the
path to be broken into a directory and a file name.

Also removed (fixed) the writing of plenary templates for machine_type
aurora_node.