Fix to allow deletion of BuildItem objects.
Something about lazy=False as part of the backrefs for BuildItem
were causing it to re-vivify while writing out plenary templates.

There was a comment there saying that it was essential but I'm not
sure how true that is anymore.  Just took it out since the current
behavior was broken.