Add general logging and out-of-band messaging facility for broker commands.
Addresses-Issue: Jira/AQUILONAQD-74
This introduces a logger variable that is passed into every render()
call.  Log messages go both into the server logs and into a global
catalog of current requests.  Logs are written into the server log
if they are higher than a specified threshold.  All logs are stored
in the catalog until the request finishes - the idea is that the
in-memory storage can be used for out of band client requests for
more information.  That out of band request can filter logs down to
a certain level.

There are two dictionaries that hold a reference to the request.
The first is cleared of the reference when the command finishes.
The second is only used if the client indicated it will be looking
at the info by sending a requestid.  That reference is only cleared
once the requestid has been queried.