aq: Enhance client to spawn a status request thread for out of band info.
The thread is spawned immediately and retried several times.  The aq
command will always try to show any relevant status info unless the
option --quiet is passed.

This aq command will specify a requestid that the broker will hold in
memory until the client has had a chance to retrieve it.  The aq
client sends the requestid both on the initial request and again when
asking for the out of band info.

The status messages are written to stderr.