Fix for running status messages against sqlite databases.
The fix is more of a hack.  When I wrote the status messages patches
I forgot that sqlite does not play well with multithreading and
parallel access.  Oops.

This introduces a --slowstatus option to the aq command to request
status messages only after the original request is processed.  It's
lame but allows the regression tests to pass when running against
sqlite.  Any dev wrappers should be updated to use either
the --slowstatus option or --quiet.  Or just run with Oracle.