Added verification test for error message when deleting a bound server.
Addresses-Issue: Jira/AQUILONAQD-119
Technically this was fixed earlier in
commit 7f39d1d01c22b041d60aa26b59de3ed3cc94915c
but I forgot that there was an incident filed for it.  This commit
adds a verification test that the ugly SQL error has been replaced
by a human-readable error message when attempting to delete a server
that is bound as a server for a service instance.