show_request: Switch to a callback based model for printing updates.
This reverses the code comments saying that locks should not be used
in the code.  They're really the best way to handle a few of the edge
cases.  This implimentation only uses the primitive Lock class (as
opposed to one of the higher level classes) and uses callbacks to
reduce the amount of thread contention.

There is around an 8.5% performance improvement on the time taken by
the regression tests.