unmap_service: Remove requirement for --archetype option.
Addresses-Issue: Jira/AQUILONAQD-246
This should have been removed when map_service was updated to not
require archetype as an option.  Enough of unmap_service changed
anyway that I took out the use of dbwrappers and replaced most of
it with calls to Base.get_unique.

Also updated show_map to bring its use of archetype in line with
the map_service and unmap_service commands.  Now archetype is a
filter on personality instead of being ignored.