Fixes for host protobuf output of ip addresses and networks.
Addresses-Issue: Jira/AQUILONAQD-254

A null IP address is now presented as an empty string instead of the
string "None".

The network id field has been removed completely as the old code was
buggy and the field is of limitted utility anyway.  Just for fun, I
fixed the bug in the code before commenting it out.