Create a logger for incoming requests and log them immediately.
Previously if a command was going to get a sqlalchemy session it
would do that first and then try to create the logger.  This
becomes a problem when all sqlalchemy threads are in use.

The original idea was that the audit id would be created in the
database and then logged.  We can revisit this decision when that
functionality is added but most likely the database id will simply
be logged at a later point (if at all).