Create a special pool of sqlalchemy sessions for non-locking actions.
Addresses-Issue: Jira/AQUILONAQD-259

These actions are generally read-only but have a new flag
(is_lock_free) for the cases where a command that can use a read-only
database connection still takes out a lock (such as put or deploy).
The db_factory refers to No Lock Sessions as opposed to Read Only ones.

The intent is to allow show, search and cat commands to have a separate
(hopefully faster-clearing) pool of database connections than the
general one that might have all of its connections waiting on
acquiring a compile lock.