allow a "generated" flag on the cat commands to see right now
The problem with the current "cat" command is that it can only
tell you about "successful" plenary files. For example, a host
template will only be left in the filesystem if it successfully
compiles, so you can't see what the plenary looks like before
a compile. This change allows "--generated" on the cat command
line which will then just give the current body() that would be
used to generate the template and goes nowhere near the filesystem