current settings acutally *lower* the max size of the pool
increasing with a few guesses for now. initial size = 30
since the various bootservers and cdb2sql in theory could
be about 15 unto themselves. Add 15 since the old pool
had a max size of 15 and was usually ok.

put the timeout in, up at 60 seconds to keep it longer than
it was, but not too much.