current settings acutally *decrease* the max size of the pool
increasing with a few guesses for now.
initial size = 30 (the bootservers and cdb2sql could possibly use 15)
Add another 15 since the old pool had a max size of 15 and was usually
sufficient.

put a slightly longer timeout (60) in place just to avoid the client
hanging forever in scripts, etc.

This is an emergency tcm, and are all temporary measures to be
addressed more thoroughly elsewhere.