Update get_unique helper to use original error description.
The current algorithm for getting a subclass when the uniqueness
constraint is on the base class involves recursion.  If there is
an error the description of the object being searched for is
generated for whatever level of the stack we're at.  This results
in a somewhat confusing error message.

The new behavior makes the error message generated be the same
regardless of the level of recursion.