deploy: Now take a target domain and by default autosync trackers.
Deploy targets are strictly domains (and strictly non-tracking
domains).  The source can be another domain or a sandbox.  Any
domains tracking the deploy target will be automatically synced
unless the nosync option is used or autosync is set to false.

Since the deploy happens in a temporary directory a dryrun option
has been added to allow people to see the results of the merge
without any consequences.