update_branch: Replace update_domain and support new fields.
Two changes of note.  I removed the ability to change ownership
information.  I also removed the ability to specify an arbitrary
pan compiler - the command can now only specify a version to use.
(The path around this version is still a configurable value.)

With these in place the command does not require aqd_admin
authorization.