Enhance the shutdown of twisted process monitoring.
The twisted facility for process monitoring includes a hook to stop
the processes that it is monitoring.  Unfortunately when this hook
is registered to fire as part of the twisted shutdown event its use
of callLater to follow SIGTERM with SIGKILL never gets a chance to
run the SIGKILL.

This change forces the SIGKILL to happen right before exiting.
Since the code was being modified anyway there is also additional
logging around the signals that are sent.