add_tor_switch: Add tor switch entries to dsdb.
Also fixes del_tor_switch to remove the dsdb entries and includes a
perl script for the upgrade to add all current switches into dsdb.
The upgrade script must be run by a user with permission to add hosts
with status aquilon to dsdb.

This will need to tweaked to handle tor_switches with an interface
named 'gigabitethernet0/1' as dsdb rejects this interface name.