Replace locking constructs with lock_queue.
Remove the compileLock and compileRelease helpers and all the random
locking mechanisms like delhost_lock.  They are replaced by the keys
defined in aquilon.server.lock.

Most of the replacements are straightforward.  The most complex
involving adding methods to the Plenary classes to indicate what
kind of locks to take out when writing or removing the associated
file.

Another change is that the commands that take out a lock specifically
to delete a system now commit the database session before releasing the
lock.  This was a further attempt to lessen deadlocks that appears to
have been unsuccessful.  It might be wise to remove these locks
entirely.

Regression tests pass.