Remove unnecessary check from 'reconfigure'
Addresses-Issue: Jira/AQUILONAQD-265
There is no need to check if 'make' has been run before 'reconfigure'.
This also makes the behavior of 'reconfigure' consistent with
'reconfigure --hostlist'.

The test from the testsuite is removed because now reconfigure would try
to call make, but the environment is not yet set up for that at this
point (no services are defined).