Changed VlanInfo to be more of a lookup table.
The old constraint was a primary key across all columns.  This
doesn't make much sense as the intent of the type field was to be
a unique qualifier on the port group.  There would not be two entries
in this table with the same vlan id and port group but with different
type entries.

However, I've gone one step further and said vlan_id is a primary key
by itself and port group has a uniqueness constraint, so entries are
one to one.