del_interface improvements
Addresses-Issue: Jira/AQUILONAQD-273
Drop the '--ip' option for del_interface since the thing the IP points
to would have to be deleted before deleting the interface anyway. If a
system is still bound to the interface then error out and print the
system's name and type in the error message.